update: ScriptReader -> ScriptParser
