+++ packages/neon-core/src/sc/index.ts
@@ -1,7 +1,7 @@
 export * from "./ContractParam";
 export * from "./core";
 export * from "./ScriptBuilder";
-export * from "./ScriptReader";
+export * from "./ScriptParser";
 export * from "./OpCode";
 export * from "./OpCodePrices";
 export * from "./InteropServiceCode";
